change error to warn (#6294)
This issue appeared while running kafka connection tests which kill the kafka broker.
When the connection to the kafka broker was lost, ERROR level logs related to CryptoOpsClientImpl appeared several times. This is because there was a failure in executing net.corda.data.crypto.wire.ops.rpc.queries.SupportedSchemesRpcQuery.

It has been verified how this is handled in the rest worker:
KeyRestResource calls listSchemes and fails at cryptoOpsClient.getSupportedSchemes.
The exception handling for listSchemes will return an InternalServerException. This extends HttpApiException and the response code will be 500.

As we return an error to the rest client and the system can recover, the log level for the CryptoOpsClient failing the operation should be WARN level instead.