CORE-20886 stop the registry from being updated when a coordinator is in status ERROR (#6365)
Numerous fixes were added to the 5.2 branch to stop Lifecyle coordinators from traversing from ERROR to UP/DOWN, however 2 use cases were missed. The lifecycle registry is the source of truth about the health of a worker and is inspected directly when the /status and /health REST endpoints are called. It is essential these report correctly when called by the k8s liveness probe to allow for unhealthy workers to be restarted.

1.) when a stop a event is received, the registry is directly updated from within the Lifecyle processor with a new status. This is eagerly blocked here when the current status is ERROR. Separately, within the registry itself any status update from ERROR is also blocked as a last resort.

2.) when a close event is received, the coordinator is removed from the registry. This could mask a worker in ERROR state by removing it from the registry. This is also blocked at the registry level.

3.) Additionally, calling start() on a coordinator in ERROR state would reset it to DOWN. This is also blocked eagerly at the LifeCycleProccessor Level.