CORE-20847 validate the flow start event size (#6394)
The FlowRestResourceImpl uses a transactional producer to send flow start events to the flow mapper. This means if the start event payload requires chunking, the producer will do so and send the event. The client will receive notification that the flow was received with a status of START_REQUESTED.
The flow mapper uses an async producer as it is more performant. It will therefore be unable to pass the message to the flow engine and it will throw a fatal exception due to trying to chunk a message with an async producer. This will cause the flow mappers topic partition to become blocked as the fatal exception will result in a retry.