CORE-20909 validate flow result cannot exceed max message size (#6393)
This PR ensures the flow result does not exceed the max allowed message size for payloads sent to the message bus. The flow result object is passed to the message bus via the FlowStatus object. If the message size is too large this will result in a fatal exception in the mediator and the flows topic partition will become blocked as the pattern will retry message after the worker restarts.