CORE-20907: stop external message api from sending messages that are too large (#6392)
Currently there is no validation to stop the user from sending a message which exceeds the max allowed message size. It has been observed that this will cause a fatal exception in the mediator message pattern. Subsequently the flow partition with the affected flow will become blocked and unable to progress.